NEW: factour out common code, add TLS tests
