+++ systemtest/src/main/java/io/strimzi/systemtest/utils/StUtils.java
@@ -448,4 +448,11 @@ public class StUtils {
         }
         return current;
     }
+
+    public static void waitForLoadBalancerService(String serviceName) {
+        LOGGER.info("Waiting when Service {} in namespace {} is ready", serviceName, kubeClient().getNamespace());
+
+        TestUtils.waitFor("service " + serviceName, Constants.POLL_INTERVAL_FOR_RESOURCE_READINESS, Constants.TIMEOUT_FOR_RESOURCE_READINESS,
+            () -> kubeClient().getClient().services().inNamespace(kubeClient().getNamespace()).withName(serviceName).get().getSpec().getExternalIPs().size() > 0);
+    }
 }
