+++ systemtest/src/test/java/io/strimzi/systemtest/HttpBridgeTlsST.java
@@ -10,13 +10,9 @@ import io.strimzi.api.kafka.model.KafkaResources;
 import io.strimzi.api.kafka.model.KafkaUser;
 import io.strimzi.api.kafka.model.listener.KafkaListenerAuthenticationTls;
 import io.strimzi.api.kafka.model.listener.KafkaListenerTls;
-import io.vertx.core.Vertx;
+import io.strimzi.systemtest.utils.StUtils;
 import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
-import io.vertx.ext.web.client.HttpResponse;
-import io.vertx.ext.web.client.WebClient;
-import io.vertx.ext.web.client.WebClientOptions;
-import io.vertx.ext.web.codec.BodyCodec;
 import io.vertx.junit5.VertxExtension;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -26,12 +22,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.TimeoutException;
 
 import static io.strimzi.systemtest.Constants.BRIDGE;
 import static io.strimzi.systemtest.Constants.REGRESSION;
@@ -44,24 +35,21 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 @Tag(BRIDGE)
 @Tag(REGRESSION)
 @ExtendWith(VertxExtension.class)
-public class HttpBridgeTlsST extends MessagingBaseST {
+public class HttpBridgeTlsST extends HttpBridgeBaseST {
     private static final Logger LOGGER = LogManager.getLogger(HttpBridgeTlsST.class);
-    public static final String NAMESPACE = "bridge-cluster-test";
-    private WebClient client;
+    public static final String NAMESPACE = "bridge-tls-cluster-test";
 
-    private String bridgeLoadBalancer = CLUSTER_NAME + "-loadbalancer";
-    private String USER_NAME = "pepa";
+    private String userName = "pepa";
+    private String bridgeHost = "";
 
     @Test
-    void testHttpBridgeSendSimpleMessage() throws Exception {
+    void testSendSimpleMessageTls() throws Exception {
         int messageCount = 50;
         String topicName = "topic-simple-send";
         // Create topic
         testClassResources.topic(CLUSTER_NAME, topicName).done();
 
         JsonObject records = generateHttpMessages(messageCount);
-        String bridgeHost = CLUSTER.client().getClient().services().inNamespace(NAMESPACE).withName(bridgeLoadBalancer).get().getSpec().getExternalIPs().get(0);
-
         JsonObject response = sendHttpRequests(records, bridgeHost, topicName);
         JsonArray offsets = response.getJsonArray("offsets");
         assertEquals(messageCount, offsets.size());
@@ -72,17 +60,16 @@ public class HttpBridgeTlsST extends MessagingBaseST {
             LOGGER.debug("offset size: {}, partition: {}, offset size: {}", offsets.size(), metadata.getInteger("partition"), metadata.getLong("offset"));
         }
 
-        receiveMessagesConsoleTls(NAMESPACE, topicName, messageCount, USER_NAME);
+        receiveMessagesConsoleTls(NAMESPACE, topicName, messageCount, userName);
     }
 
     @Test
-    void testHttpBridgeReceiveSimpleMessage() throws Exception {
+    void testReceiveSimpleMessageTls() throws Exception {
         int messageCount = 50;
         String topicName = "topic-simple-receive";
         // Create topic
         testClassResources.topic(CLUSTER_NAME, topicName).done();
 
-        String bridgeHost = CLUSTER.client().getClient().services().inNamespace(NAMESPACE).withName(bridgeLoadBalancer).get().getSpec().getExternalIPs().get(0);
         String name = "my-kafka-consumer";
         String groupId = "my-group";
 
@@ -101,7 +88,7 @@ public class HttpBridgeTlsST extends MessagingBaseST {
         // Subscribe
         assertTrue(subscribeHttpConsumer(topics, bridgeHost, Constants.HTTP_BRIDGE_DEFAULT_PORT, groupId, name));
         // Send messages to Kafka
-        sendMessagesConsoleTls(NAMESPACE, topicName, messageCount, USER_NAME);
+        sendMessagesConsoleTls(NAMESPACE, topicName, messageCount, userName);
         // Try to consume messages
         JsonArray bridgeResponse = receiveHttpRequests(bridgeHost, Constants.HTTP_BRIDGE_DEFAULT_PORT, groupId, name);
         if (bridgeResponse.size() == 0) {
@@ -114,7 +101,7 @@ public class HttpBridgeTlsST extends MessagingBaseST {
     }
 
     @BeforeAll
-    void createClassResources(Vertx vertx) {
+    void createClassResources() {
         LOGGER.info("Creating resources before the test class");
         prepareEnvForOperator(NAMESPACE);
 
@@ -143,8 +130,8 @@ public class HttpBridgeTlsST extends MessagingBaseST {
                 .endSpec().done();
 
         // Create Kafka user
-        KafkaUser userSource = testClassResources.tlsUser(CLUSTER_NAME, USER_NAME).done();
-        waitTillSecretExists(USER_NAME);
+        KafkaUser userSource = testClassResources.tlsUser(CLUSTER_NAME, userName).done();
+        waitTillSecretExists(userName);
 
         // Initialize CertSecretSource with certificate and secret names for consumer
         CertSecretSource certSecret = new CertSecretSource();
@@ -171,137 +158,7 @@ public class HttpBridgeTlsST extends MessagingBaseST {
                 .withSelector(map)
                 .endSpec().build();
         testClassResources.createServiceResource(service, NAMESPACE).done();
-        // Create http client
-        client = WebClient.create(vertx, new WebClientOptions()
-                .setSsl(false)
-                .setTrustAll(true)
-                .setVerifyHost(false));
-    }
-
-    JsonObject generateHttpMessages(int messageCount) {
-        LOGGER.info("Creating {} records for Kafka Bridge", messageCount);
-        JsonArray records = new JsonArray();
-        JsonObject json = new JsonObject();
-        for (int i = 0; i < messageCount; i++) {
-            json.put("value", "msg_" + i);
-            records.add(json);
-        }
-        JsonObject root = new JsonObject();
-        root.put("records", records);
-        return root;
-    }
-
-    JsonObject sendHttpRequests(JsonObject records, String bridgeHost, String topic) throws InterruptedException, ExecutionException, TimeoutException {
-        LOGGER.info("Sending records to Kafka Bridge");
-        CompletableFuture<JsonObject> future = new CompletableFuture<>();
-        client.post(Constants.HTTP_BRIDGE_DEFAULT_PORT, bridgeHost, "/topics/" + topic)
-                .putHeader("Content-length", String.valueOf(records.toBuffer().length()))
-                .putHeader("Content-Type", Constants.KAFKA_BRIDGE_JSON_JSON)
-                .as(BodyCodec.jsonObject())
-                .sendJsonObject(records, ar -> {
-                    if (ar.succeeded()) {
-                        LOGGER.info("Server accepted post");
-                        HttpResponse<JsonObject> response = ar.result();
-                        future.complete(response.body());
-                    } else {
-                        LOGGER.error("Server didn't accept post: {}", ar.result());
-                        future.completeExceptionally(ar.cause());
-                    }
-                });
-        return future.get(2, TimeUnit.MINUTES);
-    }
-
-    JsonArray receiveHttpRequests(String bridgeHost, int bridgePort, String groupID, String name) throws Exception {
-        CompletableFuture<JsonArray> future = new CompletableFuture<>();
-        client.get(bridgePort, bridgeHost, "/consumers/" + groupID + "/instances/" + name + "/records?timeout=" + 1000)
-                .putHeader("Accept", Constants.KAFKA_BRIDGE_JSON_JSON)
-                .as(BodyCodec.jsonArray())
-                .send(ar -> {
-                    if (ar.succeeded() && ar.result().statusCode() == 200) {
-                        HttpResponse<JsonArray> response = ar.result();
-                        if (response.body().size() > 0) {
-                            for (int i = 0; i < response.body().size(); i++) {
-                                JsonObject jsonResponse = response.body().getJsonObject(i);
-                                String kafkaTopic = jsonResponse.getString("topic");
-                                int kafkaPartition = jsonResponse.getInteger("partition");
-                                String key = jsonResponse.getString("key");
-                                int value = jsonResponse.getInteger("value");
-                                long offset = jsonResponse.getLong("offset");
-                                LOGGER.debug("Received msg: topic:{} partition:{} key:{} value:{} offset{}", kafkaTopic, kafkaPartition, key, value, offset);
-                            }
-                        } else {
-                            LOGGER.debug("Received 0 messages, going to consume again");
-                        }
-                        future.complete(response.body());
-                    } else {
-                        LOGGER.info("Cannot consume any messages!");
-                        future.completeExceptionally(ar.cause());
-                    }
-                });
-        return future.get(1, TimeUnit.MINUTES);
-    }
-
-    boolean subscribeHttpConsumer(JsonObject topics, String bridgeHost, int bridgePort, String groupId, String name) throws InterruptedException, ExecutionException, TimeoutException {
-        CompletableFuture<Boolean> future = new CompletableFuture<>();
-        client.post(bridgePort, bridgeHost,  "/consumers/" + groupId + "/instances/" + name + "/subscription")
-                .putHeader("Content-length", String.valueOf(topics.toBuffer().length()))
-                .putHeader("Content-type", Constants.KAFKA_BRIDGE_JSON)
-                .as(BodyCodec.jsonObject())
-                .sendJsonObject(topics, ar -> {
-                    if (ar.succeeded() && ar.result().statusCode() == 204) {
-                        LOGGER.info("Subscribed");
-                        future.complete(ar.succeeded());
-                    } else {
-                        LOGGER.error("Cannot subscribe consumer: {}", ar.result());
-                        future.completeExceptionally(ar.cause());
-                    }
-                });
-        return future.get(1, TimeUnit.MINUTES);
-    }
-
-    JsonObject createBridgeConsumer(JsonObject config, String bridgeHost, int bridgePort, String groupId) throws InterruptedException, ExecutionException, TimeoutException {
-        CompletableFuture<JsonObject> future = new CompletableFuture<>();
-        client.post(bridgePort, bridgeHost, "/consumers/" + groupId)
-                .putHeader("Content-length", String.valueOf(config.toBuffer().length()))
-                .putHeader("Content-type", Constants.KAFKA_BRIDGE_JSON)
-                .as(BodyCodec.jsonObject())
-                .sendJsonObject(config, ar -> {
-                    if (ar.succeeded()) {
-                        HttpResponse<JsonObject> response = ar.result();
-                        JsonObject bridgeResponse = response.body();
-                        String consumerInstanceId = bridgeResponse.getString("instance_id");
-                        String consumerBaseUri = bridgeResponse.getString("base_uri");
-                        LOGGER.debug("ConsumerInstanceId: {}", consumerInstanceId);
-                        LOGGER.debug("ConsumerBaseUri: {}", consumerBaseUri);
-                        future.complete(response.body());
-                    } else {
-                        LOGGER.error("Cannot create consumer: {}", ar.result());
-                        future.completeExceptionally(ar.cause());
-                    }
-                });
-        return future.get(1, TimeUnit.MINUTES);
-    }
-
-    boolean deleteConsumer(String bridgeHost, int bridgePort, String groupId, String name) throws InterruptedException, ExecutionException, TimeoutException {
-        LOGGER.info("Deleting consumer");
-        CompletableFuture<Boolean> future = new CompletableFuture<>();
-        client.delete(bridgePort, bridgeHost, "/consumers/" + groupId + "/instances/" + name)
-                .putHeader("Content-Type", Constants.KAFKA_BRIDGE_JSON)
-                .as(BodyCodec.jsonObject())
-                .send(ar -> {
-                    if (ar.succeeded()) {
-                        LOGGER.info("Consumer deleted");
-                        future.complete(ar.succeeded());
-                    } else {
-                        LOGGER.error("Cannot delete consumer: {}", ar.result());
-                        future.completeExceptionally(ar.cause());
-                    }
-                });
-        return future.get(1, TimeUnit.MINUTES);
-    }
-
-    @Override
-    void recreateTestEnv(String coNamespace, List<String> bindingsNamespaces) {
-        LOGGER.info("Skip env recreation after failed tests!");
+        StUtils.waitForLoadBalancerService(bridgeLoadBalancer);
+        bridgeHost = CLUSTER.client().getClient().services().inNamespace(NAMESPACE).withName(bridgeLoadBalancer).get().getSpec().getExternalIPs().get(0);
     }
 }
